chore(deps): upgrade dependencies and Gradle version
This commit updates various dependencies to their latest versions, including:
- Kotlin to 2.2.20
- Dokka to 1.8.20
- Spotless to 6.19.0
- Ktor to 3.3.0
- kotlinx-datetime to 0.7.1-0.6.x-compat
- kotlinx-coroutines-core to 1.10.2
- androidx-paging-common to 3.4.0-alpha04

It also upgrades Gradle to version 8.13.

Additionally, the following changes were made:
- Replaced Ktor CIO engine with Java engine for JVM.
- Added new targets for Linux ARM64, macOS ARM64, and MinGW x64.
- Updated iOS targets to specific architectures (Arm64, SimulatorArm64, X64).
- Applied default hierarchy template for multiplatform projects.
- Migrated to type-safe project accessors for plugins.
- Updated Kotlin compiler options and JVM target settings.
- Added `.kotlin/` to `.gitignore`.
- Added Google Maven repository to plugin management.

+++ mastodonk-core/src/commonMain/kotlin/fr/outadoc/mastodonk/api/entity/FocalPoint.kt
@@ -4,20 +4,26 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 /**
- * @see [Official docs](https://docs.joinmastodon.org/methods/statuses/media/#focal-points)
+ * Represents a focal point in an image.
+ *
+ * @see [Official docs](https://docs.joinmastodon.org/entities/MediaAttachment/#focal-points)
  */
 @Serializable
 public data class FocalPoint(
 
     /**
-     * x coordinate in [-1, +1]
+     * The x coordinate of the focal point.
+     *
+     * Float, in [-1, 1].
      */
     @SerialName("x")
-    val x: Double,
+    val x: Float,
 
     /**
-     * y coordinate in [-1, +1]
+     * The y coordinate of the focal point.
+     *
+     * Float, in [-1, 1].
      */
     @SerialName("y")
-    val y: Double
+    val y: Float
 )
